refactor: organize interfaces into root level interfaces directory and improve test robustness
- Refactored Assistants
  - Moved interfaces (`Permission`, `Model`, `ModelsResponse`, `Entry`, `EntryAnalysis`, `Message`, `ChatMessage`, `Prompt`, `ChatCompletionChoice`, `ChatCompletionUsage`, `ChatCompletionResponse`) into a new `interfaces` directory.
  - Updated import statements across modules to reference the new interface files.
  - Created separate files for each interface to promote reusability and maintainability.

- Improved Assistants test `should return the filtered model when it exists`
  - Updated test to avoid reliance on array indices in tests.
  - Modified the test to search for the expected model based on properties, making it more robust against changes in the models array.
