+++ backend/src/interfaces/assistants/Permission.ts
@@ -1,0 +1,14 @@
+export interface Permission {
+    id: string;
+    object: string;
+    created: number;
+    allow_create_engine: boolean;
+    allow_sampling: boolean;
+    allow_logprobs: boolean;
+    allow_search_indices: boolean;
+    allow_view: boolean;
+    allow_fine_tuning: boolean;
+    organization: string;
+    group?: string | null;
+    is_blocking: boolean;
+}
