+++ backend/src/interfaces/entry/Entry.ts
@@ -1,0 +1,3 @@
+export interface Entry {
+    content: string;
+}
