+++ backend/src/interfaces/entry/EntryAnalysis.ts
@@ -1,0 +1,6 @@
+import { Entry } from './Entry.js';
+
+export interface EntryAnalysis {
+    entry: Entry;
+    analysis_content: string;
+}
